requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
allennlp==2.10.1
antlr4-python3-runtime==4.8
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
backports.csv==1.0.7
base58==2.1.1
beautifulsoup4==4.12.2
black==22.1.0
bleach==6.0.0
blessed==1.20.0
blis==0.7.5
boto3==1.20.47
botocore==1.23.47
cached-path==1.1.6
cachetools==5.0.0
catalogue==2.0.6
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.11
checklist==0.0.11
cheroot==8.6.0
CherryPy==18.6.1
click==8.0.3
comm==0.1.3
commonmark==0.9.1
cryptography==36.0.1
cycler==0.11.0
cymem==2.0.6
datasets==1.18.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
docker-pycreds==0.4.0
einops==0.4.0
entrypoints==0.4
executing==1.2.0
fairscale==0.4.6
fastjsonschema==2.16.3
feedparser==6.0.8
filelock==3.4.2
flit_core==3.8.0
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.0
fsspec==2022.1.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.26
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-core==2.2.2
google-cloud-storage==1.44.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
gpustat==1.1
h5py==3.6.0
higher==0.2.1
huggingface-hub==0.10.1
hydra-core==1.1.1
idna==3.4
importlib-metadata==6.3.0
iniconfig==1.1.1
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
iso-639==0.4.5
isoduration==20.11.0
jaraco.classes==3.2.1
jaraco.collections==3.5.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.7.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.1.0
jsonnet==0.18.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.2.0
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyter-events==0.6.3
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
langcodes==3.3.0
lmdb==1.3.0
lxml==4.7.1
MarkupSafe==2.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.6
mistune==2.0.5
more-itertools==8.12.0
multidict==6.0.2
multiprocess==0.70.12.2
munch==2.5.0
murmurhash==1.0.6
mypy-extensions==0.4.3
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
nltk==3.6.5
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.22.1
nvidia-ml-py==11.525.112
omegaconf==2.1.1
packaging==23.1
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pathy==0.6.1
patternfork-nosql==3.6
pdfminer.six==20211012
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
pip==21.2.4
platformdirs==3.2.0
pluggy==1.0.0
portend==3.1.0
preshed==3.0.6
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.19.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
Pygments==2.15.0
pyparsing==3.0.7
pyrsistent==0.19.3
pytest==6.2.5
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.2
python-json-logger==2.0.7
pytz==2021.3
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
regex==2022.1.18
requests==2.30.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rsa==4.8
ruamel.yaml.clib==0.2.7
s3transfer==0.5.1
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
sentry-sdk==1.5.4
setuptools==65.6.3
sgmllib3k==1.0.0
shortuuid==1.0.8
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
stack-data==0.6.2
tempora==5.0.1
tensorboardX==2.4.1
termcolor==1.1.0
terminado==0.17.1
testpath==0.5.0
thinc==8.0.13
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.0
torch==1.10.2
torchvision==0.11.3
tornado==6.2
tqdm==4.62.3
traitlets==5.9.0
transformers==4.23.1
typer==0.4.2
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.8
wandb==0.12.10
wasabi==0.9.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
wheel==0.38.4
widgetsnbextension==4.0.7
xxhash==2.0.2
yarl==1.7.2
yaspin==2.1.0
zc.lockfile==2.0
zipp==3.15.0
